+++ src/rs/utils.rs
@@ -1,6 +1,16 @@
 /* SPDX-FileCopyrightText: © 2023 Decompollaborate */
 /* SPDX-License-Identifier: MIT */
 
+use std::{path::PathBuf, fs::File, io::Read};
+
 pub fn parse_hex(src: &str) -> u64 {
     u64::from_str_radix(src.trim_start_matches("0x"), 16).unwrap()
+}
+
+pub fn read_file_contents(file_path: &PathBuf) -> String {
+    let mut f = File::open(file_path).expect("Could not open input file");
+    let mut file_contents: String = String::new();
+    let _contents_length = f.read_to_string(&mut file_contents).expect("Not able to read the whole contents of the file");
+
+    file_contents
 }